fix(): refactor pages
