+++ webapp/src/routes/Help/index.js
@@ -8,8 +8,6 @@ import Link from '@material-ui/core/Link'
 import HttpIcon from '@material-ui/icons/Http'
 import TelegramIcon from '@material-ui/icons/Telegram'
 import GitHubIcon from '@material-ui/icons/GitHub'
-import Card from '@material-ui/core/Card'
-import CardContent from '@material-ui/core/CardContent'
 
 import styles from './styles'
 
@@ -20,58 +18,48 @@ const Help = () => {
   const { t } = useTranslation('helpRoute')
 
   return (
-    <Box>
-      <Card>
-        <CardContent>
+    <Box className={classes.helpPage}>
+      <Grid container direction="column">
+        <Grid item xs>
           <Grid container direction="column">
-            <Grid item xs>
-              <Grid container direction="column">
-                <Typography variant="h4" className={classes.title}>
-                  {t('title')}
-                </Typography>
-                <Typography variant="body2" align="justify" paragraph>
-                  {t('paragraph')}
-                </Typography>
+            <Typography className={classes.helpTitle}>{t('title')}</Typography>
+            <Typography className={classes.helpInfo}>
+              {t('paragraph')}
+            </Typography>
 
-                <Box className={classes.boxLinks}>
-                  <GitHubIcon />
-                  <Link
-                    href="https://github.com/eoscostarica/proton-affiliate"
-                    target="_blank"
-                    rel="noreferrer"
-                  >
-                    <Typography variant="body1">{t('protonGitHub')}</Typography>
-                  </Link>
-                </Box>
-                <Box className={classes.boxLinks}>
-                  <TelegramIcon />
-                  <Link
-                    href="https://t.me/protonxpr"
-                    target="_blank"
-                    rel="noreferrer"
-                  >
-                    <Typography variant="body1">
-                      {t('telegramChannel')}
-                    </Typography>
-                  </Link>
-                </Box>
-                <Box className={classes.boxLinks}>
-                  <HttpIcon />
-                  <Link
-                    href="https://www.protonchain.com/"
-                    target="_blank"
-                    rel="noreferrer"
-                  >
-                    <Typography variant="body1">
-                      {t('protonWebsite')}
-                    </Typography>
-                  </Link>
-                </Box>
-              </Grid>
-            </Grid>
+            <Box className={classes.boxLinks}>
+              <GitHubIcon />
+              <Link
+                href="https://github.com/eoscostarica/proton-affiliate"
+                target="_blank"
+                rel="noreferrer"
+              >
+                <Typography variant="body1">{t('protonGitHub')}</Typography>
+              </Link>
+            </Box>
+            <Box className={classes.boxLinks}>
+              <TelegramIcon />
+              <Link
+                href="https://t.me/protonxpr"
+                target="_blank"
+                rel="noreferrer"
+              >
+                <Typography variant="body1">{t('telegramChannel')}</Typography>
+              </Link>
+            </Box>
+            <Box className={classes.boxLinks}>
+              <HttpIcon />
+              <Link
+                href="https://www.protonchain.com/"
+                target="_blank"
+                rel="noreferrer"
+              >
+                <Typography variant="body1">{t('protonWebsite')}</Typography>
+              </Link>
+            </Box>
           </Grid>
-        </CardContent>
-      </Card>
+        </Grid>
+      </Grid>
     </Box>
   )
 }
