+++ webapp/src/routes/Help/styles.js
@@ -1,7 +1,49 @@
 export default theme => ({
-  title: {
-    marginBottom: theme.spacing(1),
-    fontWeight: 400
+  helpPage: {
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    padding: theme.spacing(2),
+    width: '100%',
+    [theme.breakpoints.up('md')]: {
+      width: 1024
+    }
+  },
+  helpTitle: {
+    fontWeight: '500',
+    fontSize: 21,
+    lineHeight: '27px',
+    textAlign: 'center',
+    letterSpacing: '0.15px',
+    color: theme.palette.common.black,
+    [theme.breakpoints.up('md')]: {
+      fontSize: 40,
+      fontWeight: 'bold',
+      fontStretch: 'normal',
+      fontStyle: 'normal',
+      lineHeight: '1.3',
+      letterSpacing: 'normal',
+      textAlign: 'left',
+      width: '100%'
+    }
+  },
+  helpInfo: {
+    fontWeight: '500',
+    margin: theme.spacing(2, 0),
+    fontSize: 16,
+    lineHeight: '24px',
+    letterSpacing: '0.44px',
+    color: '#6B717F',
+    [theme.breakpoints.up('md')]: {
+      fontSize: 21,
+      fontWeight: '500',
+      fontStretch: 'normal',
+      fontStyle: 'normal',
+      lineHeight: 'normal',
+      letterSpacing: '0.15px',
+      textAlign: 'left',
+      width: '100%'
+    }
   },
   boxLinks: {
     display: 'flex',
